pyproject.toml
[tool.poetry]
name = "ecommerce"
version = "0.1.0"
description = ""
authors = ["dopefresh <dopefresh4000@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.6"
gunicorn = "^20.1.0"
psycopg2 = "^2.9.1"
Pillow = "^8.3.1"
django-crispy-forms = "^1.12.0"
django-mathfilters = "^1.0.0"
django-countries = "^7.2.1"
whitenoise = "^5.3.0"
loguru = "^0.5.3"
django-cors-headers = "^3.8.0"
python-decouple = "^3.4"
django-allauth = "^0.45.0"
django-redis = "^5.0.0"
celery = "^5.1.2"
redis = "^3.5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
